Main interpreter: Move functionality from statusbar widget to container
That way we'll be able to emit the list of current environments to other
plugins through a signal.
